refactor: extract PDF furniture classification
